Fix tsconfig.json and rename `source-test/file.ts` to `source-test/File.ts`
